backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1